src/a2a_server/requirements.txt
a2a-sdk[http-server]==0.3.22
azure-identity==1.19.0
azure-ai-projects==2.0.0b2
openai
uvicorn[standard]==0.30.0
azure-monitor-opentelemetry==1.8.1
azure-core==1.36.0
aiohttp==3.13.2
